requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
python-pptx
